pyproject.toml
[project]
name = "tomd"
version = "0.1.0"
description = "A CLI tool to convert PDF, DOC, and Excel documents into Markdown format"
authors = [{name = "Antigravity", email = "antigravity@example.com"}]
requires-python = ">=3.9"
dependencies = [
    "click>=8.1.0",
    "python-docx>=1.1.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "tabulate>=0.9.0",
    "pymupdf4llm>=0.0.5",
]

[project.scripts]
tomd = "tomd.cli:cli"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
